requirements.txt
langchain
openai
chromadb
unstructured
python-dotenv
tiktoken
